Changed positioning
Had to fork it, because no matter what I tried, adding this in the options during creation didn't work. For other, similar, issues, the author also suggests forking the plugin. So here's my lightweight fork that only (for now) moves the cookie bar to the bottom of the screen by default.